Closed thread after disconnecting
- One issue solved (closing thread after disconnecting)
- I tried to work on setting player's score on the server side (since the score is locally saved only in SinglePlayerGameCtrl class), created some APIs and I;ll work my way through later.

+++ commons/src/main/java/commons/player/SimpleUser.java
@@ -60,7 +60,7 @@ public class SimpleUser {
         this.id = id;
     }
 
-    public long getScore() {
+    public int getScore() {
         return score;
     }
